requirements.txt
django~=3.2.10
django-autoslug==1.9.8
django-model-utils==4.1.0
pillow==8.0.1
django-debug-toolbar==3.1.1

# factories
factory-boy==3.1.0

# pytest
pytest==6.1.2
pytest-django==4.1.0

# widget_tweaks
django-widget-tweaks==1.4.8

# django-crispy
django-crispy-forms==1.11.0
django-localflavor==3.0.1

# allauth
django-allauth==0.50.0

# payment
mercadopago==2.0.6
django-environ==0.4.5
pytest-mock==3.5.1
requests==2.25.1